Define API endpoint to verify mobile numbers
The page that prompts new users to enter their mobile number
was changed. It now includes a JavaScript file that sends an
AJAX POST request with the user's mobile number. When the
server receives the mobile number, it sends the user an SMS
message with a unique 6-digit code that they must enter.