Change create user form
The form that is used to create new users didn't previously
accept an email address as one of the form fields. This caused
all new user objects to have their email attribute assigned to
the empty string ('') by default. However, the email attribute is
required to be unique. This meant that the site wouldn't allow
more than one user to be created.

This has been rectified, and now the initial page that prompts
users to enter information when creating a user account requires
their email address. As a result, the page that prompts them for
their email address has been removed, because it is redundant.